Refactors Constant, ConstantCollection, Equation and EquationCollection into statement.py module
