+++ pyTrnsysType/input_file.py
@@ -1,4 +1,3 @@
-import collections
 import itertools
 import logging as lg
 import re
@@ -7,13 +6,10 @@ import tempfile
 import tabulate
 from path import Path
 from shapely.geometry import LineString, Point
-from sympy import Expr, Symbol
 
 from pyTrnsysType import (
-    TypeVariable,
     TrnsysModel,
     Component,
-    StudioHeader,
     MetaData,
     AnchorPoint,
     ComponentCollection,
@@ -34,16 +30,14 @@ from pyTrnsysType.statements import (
     EqSolver,
     End,
     Solver,
-    Statement,
     Width,
+    Constant,
+    ConstantCollection,
+    Equation,
+    EquationCollection,
 )
-from pyTrnsysType.utils import print_my_latex, TypeVariableSymbol, get_rgb_from_int
-from .trnsymodel import (
-    ParameterCollection,
-    InputCollection,
-    ExternalFileCollection,
-    _studio_to_linestyle,
-)
+from pyTrnsysType.utils import get_rgb_from_int
+from .trnsymodel import _studio_to_linestyle
 
 
 class Name(object):
@@ -100,580 +94,6 @@ class Trace:
 class Format:
     # Todo: Implement Format
     pass
-
-
-class Constant(Statement):
-    """The CONSTANTS statement is useful when simulating a number of systems
-    with identical component configurations but with different parameter values,
-    initial input values, or initial values of time dependent variables.
-    """
-
-    _new_id = itertools.count(start=1)
-    instances = {}
-
-    def __init__(self, name=None, equals_to=None, doc=None):
-        """
-        Args:
-            name (str): The left hand side of the equation.
-            equals_to (str, TypeVariable): The right hand side of the equation.
-            doc (str, optional): A small description optionally printed in the
-                deck file.
-        """
-        super().__init__()
-        try:
-            c_ = Constant.instances[name]
-        except:
-            self._n = next(self._new_id)
-            self.name = name
-            self.equals_to = equals_to
-            self.doc = doc
-        else:
-            self._n = c_._n
-            self.name = c_.name
-            self.equals_to = c_.equals_to
-            self.doc = c_.doc
-        finally:
-            Constant.instances.update({self.name: self})
-
-    @classmethod
-    def from_expression(cls, expression, doc=None):
-        """Create a Constant from a string expression. Anything before the equal
-        sign ("=") will become the Constant's name and anything after will
-        become the equality statement.
-
-        Hint:
-            The simple expressions are processed much as FORTRAN arithmetic
-            statements are, with one significant exceptions. Expressions are
-            evaluated from left to right with no precedence accorded to any
-            operation over another. This rule must constantly be borne in mind
-            when writing long expressions.
-
-        Args:
-            expression (str): A user-defined expression to parse.
-            doc (str, optional): A small description optionally printed in the
-                deck file.
-        """
-        if "=" not in expression:
-            raise ValueError(
-                "The from_expression constructor must contain an expression "
-                "with the equal sign"
-            )
-        a, b = expression.split("=")
-        return cls(a.strip(), b.strip(), doc=doc)
-
-    @property
-    def constant_number(self):
-        """The equation number. Unique"""
-        return self._n
-
-    def _to_deck(self):
-        return self.equals_to
-
-
-class ConstantCollection(collections.UserDict, Component):
-    """A class that behaves like a dict and that collects one or more
-    :class:`Constants`.
-
-    You can pass a dict of Equation or you can pass a list of Equation. In
-    the latter, the :attr:`Equation.name` attribute will be used as a key.
-    """
-
-    def __init__(self, mutable=None, name=None):
-        """Initialize a new ConstantCollection.
-
-        Example:
-            >>> c_1 = Constant.from_expression("A = 1")
-            >>> c_2 = Constant.from_expression("B = 2")
-            >>> ConstantCollection([c_1, c_2])
-
-        Args:
-            mutable (Iterable, optional): An iterable.
-            name (str): A user defined name for this collection of constants.
-                This name will be used to identify this block of constants in
-                the .dck file;
-        """
-        if isinstance(mutable, list):
-            _dict = {f.name: f for f in mutable}
-        else:
-            _dict = mutable
-        super().__init__(_dict)
-        self.name = Name(name)
-        self.studio = StudioHeader.from_component(self)
-        self._unit = next(TrnsysModel.new_id)
-
-    def __getitem__(self, key):
-        """
-        Args:
-            key:
-        """
-        value = super().__getitem__(key)
-        return value
-
-    def __repr__(self):
-        return self._to_deck()
-
-    def __hash__(self):
-        return hash(str(self))
-
-    def __eq__(self, other):
-        return hash(self) == hash(other)
-
-    def update(self, E=None, **F):
-        """D.update([E, ]**F). Update D from a dict/list/iterable E and F.
-        If E is present and has a .keys() method, then does:  for k in E: D[
-        k] = E[k]
-        If E is present and lacks a .keys() method, then does:  for cts.name,
-        cts in E: D[cts.name] = cts
-        In either case, this is followed by: for k in F:  D[k] = F[k]
-
-        Args:
-            E (list, dict or Constant): The constant to add or update in D (
-                self).
-            F (list, dict or Constant): Other constants to update are passed.
-        """
-        if isinstance(E, Constant):
-            E.model = self
-            _e = {E.name: E}
-        elif isinstance(E, list):
-            _e = {cts.name: cts for cts in E}
-        else:
-            for v in E.values():
-                if not isinstance(v, Constant):
-                    raise TypeError(
-                        "Can only update an ConstantCollection with a"
-                        "Constant, not a {}".format(type(v))
-                    )
-            _e = {v.name: v for v in E.values()}
-        k: Constant
-        for k in F:
-            if isinstance(F[k], dict):
-                _f = {v.name: v for k, v in F.items()}
-            elif isinstance(F[k], list):
-                _f = {cts.name: cts for cts in F[k]}
-            else:
-                raise TypeError(
-                    "Can only update an ConstantCollection with a"
-                    "Constant, not a {}".format(type(F[k]))
-                )
-            _e.update(_f)
-        super(ConstantCollection, self).update(_e)
-
-    @property
-    def size(self):
-        return len(self)
-
-    @property
-    def unit_number(self):
-        return self._unit
-
-    def _to_deck(self):
-        """To deck representation
-
-        Examples::
-
-            CONSTANTS n
-            NAME1 = ... constant 1 ...
-            NAME2 = ... constant 2 ...
-            •
-            •
-            •
-            NAMEn = ... constant n ...
-        """
-        header_comment = '* CONSTANTS "{}"\n\n'.format(self.name)
-        head = "CONSTANTS {}\n".format(len(self))
-        v_ = ((equa.name, "=", str(equa)) for equa in self.values())
-        core = tabulate.tabulate(v_, tablefmt="plain", numalign="left")
-        return str(header_comment) + str(head) + str(core)
-
-    def _get_inputs(self):
-        """inputs getter. Sorts by order number each time it is called
-        """
-        return self
-
-    def _get_outputs(self):
-        """outputs getter. Since self is already a  dict, return self.
-        """
-        return self
-
-
-class Equation(Statement):
-    """The EQUATIONS statement allows variables to be defined as algebraic
-    functions of constants, previously defined variables, and outputs from
-    TRNSYS components. These variables can then be used in place of numbers in
-    the TRNSYS input file to represent inputs to components; numerical values of
-    parameters; and initial values of inputs and time-dependent variables. The
-    capabilities of the EQUATIONS statement overlap but greatly exceed those of
-    the CONSTANTS statement described in the previous section.
-
-    Hint:
-        In pyTrnsysType, the Equation class works hand in hand with the
-        :class:`EquationCollection` class. This class behaves a little bit like
-        the equation component in the TRNSYS Studio, meaning that you can list
-        equation in a block, give it a name, etc. See the
-        :class:`EquationCollection` class for more details.
-    """
-
-    _new_id = itertools.count(start=1)
-
-    def __init__(self, name=None, equals_to=None, doc=None, model=None):
-        """
-        Args:
-            name (str): The left hand side of the equation.
-            equals_to (str, TypeVariable): The right hand side of the equation.
-            doc (str, optional): A small description optionally printed in the
-                deck file.
-        """
-        super().__init__()
-        self._n = next(self._new_id)
-        self.name = name
-        self.equals_to = equals_to
-        self.doc = doc
-        self.model = model  # the TrnsysModel this Equation belongs to.
-
-        self._connected_to = None
-
-    def __repr__(self):
-        return " = ".join([self.name, self._to_deck()])
-
-    def __str__(self):
-        return self.__repr__()
-
-    @classmethod
-    def from_expression(cls, expression, doc=None):
-        """Create an equation from a string expression. Anything before the
-        equal sign ("=") will become a Constant and anything after will become
-        the equality statement.
-
-        Example:
-            Create a simple expression like so:
-
-            >>> equa1 = Equation.from_expression("TdbAmb = [011,001]")
-
-        Args:
-            expression (str): A user-defined expression to parse.
-            doc (str, optional): A small description optionally printed in the
-                deck file.
-        """
-        if "=" not in expression:
-            raise ValueError(
-                "The from_expression constructor must contain an expression "
-                "with the equal sign"
-            )
-        a, b = expression.split("=")
-        return cls(a.strip(), b.strip(), doc=doc)
-
-    @classmethod
-    def from_symbolic_expression(cls, name, exp, *args, doc=None):
-        """Crate an equation with a combination of a generic expression (with
-        placeholder variables) and a list of arguments. The underlying engine
-        will use Sympy and symbolic variables. You can use a mixture of
-        :class:`TypeVariable` and :class:`Equation`, :class:`Constant` as
-        well as the python default :class:`str`.
-
-        .. Important::
-
-            If a `str` is passed in place of an expression argument (
-            :attr:`args`), make sure to declare that string as an Equation or
-            a Constant later in the routine.
-
-        Examples:
-            In this example, we define a variable (var_a) and we want it to be
-            equal to the 'Outlet Air Humidity Ratio' divided by 12 + log(
-            Temperature to heat source). In a TRNSYS deck file one would have to
-            manually determine the unit numbers and output numbers and write
-            something like : '[1, 2]/12 + log([1, 1])'. With the
-            :func:`~from_symbolic_expression`, we can do this very simply:
-
-            1. first, define the name of the variable:
-
-            >>> name = "var_a"
-
-            2. then, define the expression as a string. Here, the variables `a`
-            and `b` are symbols that represent the two type outputs. Note that
-            their name has bee chosen arbitrarily.
-
-            >>> exp = "log(a) + b / 12"
-            >>> # would be also equivalent to
-            >>> exp = "log(x) + y / 12"
-
-            3. here, we define the actual variables (the type outputs) after
-            loading our model from its proforma:
-
-            >>> from pyTrnsysType import TrnsysModel
-            >>> fan = TrnsysModel.from_xml("fan_type.xml")
-            >>> vars = (fan.outputs[0], fan.outputs[1])
-
-            .. Important::
-
-                The order of the symbolic variable encountered in the string
-                expression (step 2), from left to right, must be the same for
-                the tuple of variables. For instance, `a` is followed by `b`,
-                therefore `fan.outputs[0]` is followed by `fan.outputs[1]`.
-
-            4. finally, we create the Equation. Note that vars is passed with
-            the '*' declaration to unpack the tuple.
-
-            >>> from pyTrnsysType.input_file import Equation
-            >>> eq = Equation.from_symbolic_expression(name, exp, *vars)
-            >>> print(eq)
-            [1, 1]/12 + log([1, 2])
-
-        Args:
-            name (str): The name of the variable (left-hand side), of the
-                equation.
-            exp (str): The expression to evaluate. Use any variable name and
-                mathematical expression.
-            *args (tuple): A tuple of :class:`TypeVariable` that will replace
-                the any variable name specified in the above expression.
-            doc (str, optional): A small description optionally printed in the
-                deck file.
-
-        Returns:
-            Equation: The Equation Statement object.
-        """
-        from sympy.parsing.sympy_parser import parse_expr
-
-        exp = parse_expr(exp)
-
-        if len(exp.free_symbols) != len(args):
-            raise AttributeError(
-                "The expression does not have the same number of "
-                "variables as arguments passed to the symbolic expression "
-                "parser."
-            )
-        for i, arg in enumerate(sorted(exp.free_symbols, key=lambda sym: sym.name)):
-            new_symbol = args[i]
-            if isinstance(new_symbol, TypeVariable):
-                exp = exp.subs(arg, TypeVariableSymbol(new_symbol))
-            elif isinstance(new_symbol, (Equation, Constant)):
-                exp = exp.subs(arg, Symbol(new_symbol.name))
-            else:
-                exp = exp.subs(arg, Symbol(new_symbol))
-        return cls(name, exp)
-
-    @property
-    def eq_number(self):
-        """The equation number. Unique"""
-        return self._n
-
-    @property
-    def idx(self):
-        """The 0-based index of the Equation"""
-        ns = {e: i for i, e in enumerate(self.model)}
-        return ns[self.name]
-
-    @property
-    def one_based_idx(self):
-        """The 1-based index of the Equation"""
-        return self.idx + 1
-
-    @property
-    def unit_number(self):
-        return self.model.unit_number
-
-    @property
-    def is_connected(self):
-        """Whether or not this TypeVariable is connected to another type"""
-        return self.connected_to is not None
-
-    @property
-    def connected_to(self):
-        """The TrnsysModel to which this component is connected"""
-        return self._connected_to
-
-    def _to_deck(self):
-        if isinstance(self.equals_to, TypeVariable):
-            return "[{unit_number}, {output_id}]".format(
-                unit_number=self.equals_to.model.unit_number,
-                output_id=self.equals_to.one_based_idx,
-            )
-        elif isinstance(self.equals_to, Expr):
-            return print_my_latex(self.equals_to)
-        else:
-            return self.equals_to
-
-
-class EquationCollection(collections.UserDict, Component):
-    """A class that behaves like a dict and that collects one or more
-    :class:`Equations`. This class behaves a little bit like the equation
-    component in the TRNSYS Studio, meaning that you can list equation in a
-    block, give it a name, etc.
-
-    You can pass a dict of Equation or you can pass a list of Equation. In
-    this case, the :attr:`Equation.name` attribute will be used as a key.
-
-    Hint:
-        Creating equations in PyTrnsysType is done trough the :class:`Equation`
-        class. Equations are than collected in this EquationCollection. See the
-        :class:`Equation` class for more details.
-    """
-
-    def __init__(self, mutable=None, name=None):
-        """Initialize a new EquationCollection.
-
-        Example:
-            >>> equa1 = Equation.from_expression("TdbAmb = [011,001]")
-            >>> equa2 = Equation.from_expression("rhAmb = [011,007]")
-            >>> EquationCollection([equa1, equa2])
-
-        Args:
-            mutable (Iterable, optional): An iterable (dict or list).
-            name (str): A user defined name for this collection of equations.
-                This name will be used to identify this block of equations in
-                the .dck file;
-        """
-        if isinstance(mutable, list):
-            _dict = {f.name: f for f in mutable}
-        else:
-            _dict = mutable
-        super().__init__(_dict)
-        self.name = Name(name)
-        self._unit = next(TrnsysModel.new_id)
-        self.studio = StudioHeader.from_component(self)
-
-    def __getitem__(self, key):
-        """
-        Args:
-            key:
-        """
-        if isinstance(key, int):
-            value = list(self.data.values())[key]
-        else:
-            value = super().__getitem__(key)
-        return value
-
-    # def __hash__(self):
-    #     return self.unit_number
-
-    def __repr__(self):
-        return self._to_deck()
-
-    def __setitem__(self, key, value):
-        # optional processing here
-        value.model = self
-        super().__setitem__(key, value)
-
-    def __hash__(self):
-        return self._unit
-
-    def update(self, E=None, **F):
-        """D.update([E, ]**F). Update D from a dict/list/iterable E and F.
-        If E is present and has a .keys() method, then does:  for k in E: D[
-        k] = E[k]
-        If E is present and lacks a .keys() method, then does:  for eq.name,
-        eq in E: D[eq.name] = eq
-        In either case, this is followed by: for k in F:  D[k] = F[k]
-
-        Args:
-            E (list, dict or Equation): The equation to add or update in D (
-                self).
-            F (list, dict or Equation): Other Equations to update are passed.
-
-        Returns:
-            None
-        """
-        if isinstance(E, Equation):
-            E.model = self
-            _e = {E.name: E}
-        elif isinstance(E, list):
-            _e = {eq.name: eq for eq in E}
-        else:
-            for v in E.values():
-                if not isinstance(v, Equation):
-                    raise TypeError(
-                        "Can only update an EquationCollection with an"
-                        "Equation, not a {}".format(type(v))
-                    )
-            _e = {v.name: v for v in E.values()}
-        k: Equation
-        for k in F:
-            if isinstance(F[k], dict):
-                _f = {v.name: v for k, v in F.items()}
-            elif isinstance(F[k], list):
-                _f = {eq.name: eq for eq in F[k]}
-            else:
-                raise TypeError(
-                    "Can only update an EquationCollection with an"
-                    "Equation, not a {}".format(type(F[k]))
-                )
-            _e.update(_f)
-        super(EquationCollection, self).update(_e)
-
-    def setdefault(self, key, value=None):
-        if key not in self:
-            self[key] = value
-        return self[key]
-
-    @property
-    def size(self):
-        return len(self)
-
-    @property
-    def unit_number(self):
-        return self._unit
-
-    @property
-    def unit_name(self):
-        """This type does not have a unit_name. Return component name"""
-        return self.name
-
-    @property
-    def model(self):
-        """This model does not have a proforma. Return class name."""
-        return self.__class__.__name__
-
-    def _to_deck(self):
-        """To deck representation
-
-        Examples::
-
-            EQUATIONS n
-            NAME1 = ... equation 1 ...
-            NAME2 = ... equation 2 ...
-            •
-            •
-            •
-            NAMEn = ... equation n ...
-        """
-        header_comment = '* EQUATIONS "{}"\n\n'.format(self.name)
-        head = "EQUATIONS {}\n".format(len(self))
-        v_ = ((equa.name, "=", equa._to_deck()) for equa in self.values())
-        core = tabulate.tabulate(v_, tablefmt="plain", numalign="left")
-        return str(header_comment) + str(head) + str(core)
-
-    def _get_inputs(self):
-        """inputs getter. Sorts by order number each time it is called
-        """
-        return self
-
-    def _get_outputs(self):
-        """outputs getter. Since self is already a  dict, return self.
-        """
-        return self
-
-    def _get_ordered_filtered_types(self, classe_, store):
-        """
-        Args:
-            classe_:
-            store:
-        """
-        return collections.OrderedDict(
-            (attr, self._meta[store][attr])
-            for attr in sorted(
-                self._get_filtered_types(classe_, store),
-                key=lambda key: self._meta[store][key].order,
-            )
-        )
-
-    def _get_filtered_types(self, classe_, store):
-        """
-        Args:
-            classe_:
-            store:
-        """
-        return filter(
-            lambda kv: isinstance(self._meta[store][kv], classe_), self._meta[store]
-        )
 
 
 class ControlCards(object):
