+++ tests/test_xml.py
@@ -291,7 +291,8 @@ class TestTrnsysModel:
             fan_type.connect_to(fan_2, mapping=mapping)
 
     def test_connect_eqcollection_to_type(self, fan_type, tank_type):
-        from pyTrnsysType import Equation, EquationCollection
+        from pyTrnsysType.input_file import EquationCollection
+        from pyTrnsysType.input_file import Equation
 
         ec = EquationCollection(Equation.from_expression("my=b"))
         ec.set_canvas_position((500, 500))
@@ -350,8 +351,7 @@ class TestTrnsysModel:
         assert fan_type.studio.position == Point(500, 400)
 
     def test_set_link_style_to_itself(self, fan_type):
-        with pytest.raises(NotImplementedError):
-            fan_type.set_link_style(fan_type)
+        fan_type.set_link_style(fan_type)
 
     def test_set_link_style(self, fan_type):
         fan2 = fan_type.copy()
@@ -508,7 +508,8 @@ class TestConstantsAndEquations:
 
     @pytest.fixture()
     def equation_block(self):
-        from pyTrnsysType.input_file import Equation, EquationCollection
+        from pyTrnsysType.input_file import EquationCollection
+        from pyTrnsysType.input_file import Equation
 
         equa1 = Equation.from_expression("TdbAmb = [011,001]")
         equa2 = Equation.from_expression("rhAmb = [011,007]")
@@ -529,7 +530,8 @@ class TestConstantsAndEquations:
 
     @pytest.fixture()
     def constant_block(self):
-        from pyTrnsysType.input_file import Constant, ConstantCollection
+        from pyTrnsysType.input_file import ConstantCollection
+        from pyTrnsysType.input_file import Constant
 
         c_1 = Constant.from_expression("A = 1", doc="A is a constant")
         c_2 = Constant.from_expression("B = 2")
@@ -543,7 +545,7 @@ class TestConstantsAndEquations:
         assert equation_block.unit_number > 0
 
     def test_symbolic_expression(self, tank_type, fan_type):
-        from pyTrnsysType import Constant
+        from pyTrnsysType.input_file import Constant
 
         name = "var_a"
         exp = "log(a) + (b / 12 - c) * d"
@@ -558,7 +560,7 @@ class TestConstantsAndEquations:
 
     def test_malformed_symbolic_expression(self, tank_type, fan_type):
         """passed only two args while expression asks for 3"""
-        from pyTrnsysType import Constant
+        from pyTrnsysType.input_file import Constant
 
         name = "var_a"
         exp = "log(a) + b / 12 - c"
@@ -575,7 +577,8 @@ class TestConstantsAndEquations:
         eq = EquationCollection()
 
     def test_equation_collection(self, equation_block):
-        from pyTrnsysType.input_file import Equation, EquationCollection
+        from pyTrnsysType.input_file import EquationCollection
+        from pyTrnsysType.input_file import Equation
 
         equa_col_2 = EquationCollection(
             [equa for equa in equation_block.values()], name="test"
@@ -591,7 +594,8 @@ class TestConstantsAndEquations:
 
     def test_update_equation_collection(self, equation_block):
         """test different .update() recipes"""
-        from pyTrnsysType.input_file import Equation, EquationCollection
+        from pyTrnsysType.input_file import EquationCollection
+        from pyTrnsysType.input_file import Equation
 
         list_equations = [equa for equa in equation_block.values()]
 
@@ -618,7 +622,8 @@ class TestConstantsAndEquations:
         ec.update(list_equations, F=list_equations)
 
     def test_equation_with_typevariable(self, fan_type):
-        from pyTrnsysType.input_file import Equation, EquationCollection
+        from pyTrnsysType.input_file import EquationCollection
+        from pyTrnsysType.input_file import Equation
 
         fan_type._unit = 1
         equa1 = Equation("T_out", fan_type.outputs[0])
@@ -630,7 +635,8 @@ class TestConstantsAndEquations:
 
     def test_two_unnamed_equationcollection(self, fan_type):
         """make sure objects with same name=None can be created"""
-        from pyTrnsysType.input_file import Equation, EquationCollection
+        from pyTrnsysType.input_file import EquationCollection
+        from pyTrnsysType.input_file import Equation
 
         eq1 = Equation("A", fan_type.outputs[0])
         EquationCollection([eq1])
@@ -650,7 +656,8 @@ class TestConstantsAndEquations:
 
     def test_update_constant_collection(self, constant_block):
         """test different .update() recipes"""
-        from pyTrnsysType.input_file import Constant, ConstantCollection
+        from pyTrnsysType.input_file import ConstantCollection
+        from pyTrnsysType.input_file import Constant
 
         list_constants = [cts for cts in constant_block.values()]
 
@@ -680,16 +687,18 @@ class TestConstantsAndEquations:
 class TestDeck:
     @pytest.fixture(scope="class")
     def pvt_deck(self):
-        from pyTrnsysType import Deck
+        from pyTrnsysType.input_file import Deck
 
         file = "tests/input_files/test_deck.dck"
         with patch("builtins.input", return_value="y"):
             dck = Deck.from_file(file, proforma_root="tests/input_files")
             yield dck
 
+    @pytest.mark.xfail(raises=ValueError)
     @pytest.fixture(scope="class")
     def irregular_deck(self):
-        from pyTrnsysType import Deck
+        from pyTrnsysType.input_file import Deck
+
         file = "tests/input_files/Case600h10.dck"
         with patch("builtins.input", return_value="y"):
             dck = Deck.from_file(file)
@@ -699,11 +708,12 @@ class TestDeck:
     def G(self, pvt_deck):
         yield pvt_deck.graph
 
+    @pytest.mark.xfail(raises=ValueError)
     def test_irragular_dekc(self, irregular_deck):
         assert irregular_deck
 
     def test_update_with_model(self, weather_type, tank_type, pipe_type):
-        from pyTrnsysType import Deck, ControlCards
+        from pyTrnsysType.input_file import Deck, ControlCards
 
         cc = ControlCards.basic_template()
         dck = Deck("test", cc)
